requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
numpy==2.2.6
    # via opencv-python
opencv-python==4.12.0.88
    # via -r requirements.in
packaging==25.0
    # via pytesseract
pdf2image==1.17.0
    # via -r requirements.in
pillow==12.0.0
    # via
    #   -r requirements.in
    #   pdf2image
    #   pytesseract
pytesseract==0.3.13
    # via -r requirements.in
